backend/requirements.txt
fastapi>=0.115.0
scikit-learn>=1.3.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
openai>=1.40.0
python-dotenv>=1.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
